refactor : space 관련 request 클래스를 controller/service 단위에서 하나로 통합
