+++ src/main/java/com/moim/backend/domain/space/request/CreateSpaceCalendarRequest.java
@@ -1,9 +1,10 @@
-package com.moim.backend.domain.space.request.controller;
+package com.moim.backend.domain.space.request;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateTimeDeserializer;
 import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
+import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -12,17 +13,28 @@ import org.springframework.format.annotation.DateTimeFormat;
 
 import java.time.LocalDateTime;
 
+@Getter
 @AllArgsConstructor
 @NoArgsConstructor
-@Getter
-public class SpaceTimeLineRequest {
+public class CreateSpaceCalendarRequest {
 
     @NotNull
     private Long spaceId;
+
+    @NotBlank(message = "스케줄명을 입력하지 않았습니다.")
+    @NotNull
+    private String scheduleName;
 
     @DateTimeFormat(pattern = "yyyy-MM-dd'T'HH:mm:ss")
     @JsonSerialize(using = LocalDateTimeSerializer.class)
     @JsonDeserialize(using = LocalDateTimeDeserializer.class)
     @NotNull(message = "날짜 및 시간을 입력하지 않았습니다.")
     private LocalDateTime date;
+
+    @NotBlank(message = "메모를 입력하지 않았습니다.")
+    @NotNull
+    private String note;
+
+    private String locationName;
+
 }
