+++ src/main/java/com/moim/backend/domain/space/request/SpaceNameUpdateRequest.java
@@ -1,6 +1,5 @@
-package com.moim.backend.domain.space.request.controller;
+package com.moim.backend.domain.space.request;
 
-import com.moim.backend.domain.space.request.service.SpaceNameUpdateServiceRequest;
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -12,10 +11,4 @@ import lombok.NoArgsConstructor;
 public class SpaceNameUpdateRequest {
     @NotBlank(message = "수정할 그룹 이름을 입력하지 않았습니다.")
     private String groupName;
-
-    public SpaceNameUpdateServiceRequest toServiceRequest() {
-        return SpaceNameUpdateServiceRequest.builder()
-                .groupName(groupName)
-                .build();
-    }
 }
