+++ src/main/java/com/moim/backend/domain/space/service/SpaceCalendarService.java
@@ -4,11 +4,10 @@ import com.moim.backend.domain.space.entity.Space;
 import com.moim.backend.domain.space.entity.SpaceCalendar;
 import com.moim.backend.domain.space.repository.SpaceCalendarRepository;
 import com.moim.backend.domain.space.repository.SpaceRepository;
-import com.moim.backend.domain.space.request.controller.CreateSpaceCalendarRequest;
-import com.moim.backend.domain.space.request.controller.SpaceTimeLineRequest;
+import com.moim.backend.domain.space.request.CreateSpaceCalendarRequest;
+import com.moim.backend.domain.space.request.SpaceTimeLineRequest;
 import com.moim.backend.domain.space.response.space.CreateSpaceCalendarResponse;
 import com.moim.backend.domain.space.response.space.SpaceTimeLineResponse;
-import com.moim.backend.domain.user.entity.Users;
 import com.moim.backend.global.common.exception.CustomException;
 import com.moim.backend.global.util.DateParser;
 import lombok.RequiredArgsConstructor;
