+++ src/test/java/com/moim/backend/docs/space/SpaceControllerDocsTest.java
@@ -3,10 +3,10 @@ package com.moim.backend.docs.space;
 import com.epages.restdocs.apispec.ResourceSnippetParameters;
 import com.moim.backend.RestDocsSupport;
 import com.moim.backend.domain.space.controller.SpaceController;
-import com.moim.backend.domain.space.request.controller.SpaceCreateRequest;
-import com.moim.backend.domain.space.request.controller.SpaceNameUpdateRequest;
-import com.moim.backend.domain.space.request.controller.SpaceParticipateRequest;
-import com.moim.backend.domain.space.request.controller.SpaceParticipateUpdateRequest;
+import com.moim.backend.domain.space.request.SpaceCreateRequest;
+import com.moim.backend.domain.space.request.SpaceNameUpdateRequest;
+import com.moim.backend.domain.space.request.SpaceParticipateRequest;
+import com.moim.backend.domain.space.request.SpaceParticipateUpdateRequest;
 import com.moim.backend.domain.space.response.*;
 import com.moim.backend.domain.space.response.space.*;
 import com.moim.backend.domain.space.service.SpaceService;
