+++ src/test/java/com/moim/backend/domain/space/service/SpaceCalendarServiceTest.java
@@ -7,14 +7,13 @@ import com.moim.backend.domain.space.entity.TransportationType;
 import com.moim.backend.domain.space.repository.ParticipationRepository;
 import com.moim.backend.domain.space.repository.SpaceCalendarRepository;
 import com.moim.backend.domain.space.repository.SpaceRepository;
-import com.moim.backend.domain.space.request.controller.CreateSpaceCalendarRequest;
-import com.moim.backend.domain.space.request.controller.SpaceTimeLineRequest;
+import com.moim.backend.domain.space.request.CreateSpaceCalendarRequest;
+import com.moim.backend.domain.space.request.SpaceTimeLineRequest;
 import com.moim.backend.domain.space.response.space.CreateSpaceCalendarResponse;
 import com.moim.backend.domain.space.response.space.SpaceTimeLineResponse;
 import com.moim.backend.domain.user.entity.Users;
 import com.moim.backend.domain.user.repository.UserRepository;
 import jakarta.persistence.EntityManager;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,14 +23,11 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.format.TextStyle;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 
 import static com.moim.backend.domain.space.entity.TransportationType.*;
 import static com.moim.backend.domain.space.response.space.SpaceTimeLineResponse.*;
-import static java.time.format.TextStyle.*;
 import static java.time.format.TextStyle.SHORT;
 import static java.util.Locale.*;
 import static org.assertj.core.api.Assertions.*;
